Extract UpgradeWaitingView
